\docType{data}
\name{sotiriou03.data.GE}
\alias{sotiriou03.data.GE}
\title{Gene expression data of patients with breast cancer}
\format{a \code{matrix} with 99 rows and 7650 columns}
\description{
  This data set contains the gene expression of 7650 probes
  measured with cDNA microarrays, for 99 patients with
  breast cancer. It was published by Sotiriou and collegues
  in 2003.
}
\references{
  Sotiriou's PNAS paper ... complete ref
}
\keyword{datasets}

