\docType{data}
\name{sotiriou03.descr}
\alias{sotiriou03.descr}
\title{Clinical data associated to the Sotoriou's breast cancer data set.}
\format{a \code{data.frame} with 99 rows and 31 columns}
\source{
  Sotiriou's PNAS paper ... complete ref
}
\description{
  This data set contains the clinical information for the
  99 patients with breast cancer included in the study by
  Sotiriou and collegues.
}
\keyword{datasets}

