\docType{data}
\name{sotiriou03.geneID}
\alias{sotiriou03.geneID}
\title{Annotation of the genes included in the Sotoriou's breast cancer data set.}
\format{a \code{data.frame} with 7650 rows and 3 columns}
\source{
  Sotiriou's PNAS paper ... complete ref
}
\description{
  This data set contains the annotation of the 7650 probes
  included in the study by Sotiriou and collegues. It
  contains the clone ID, the gene names and the unique ID
  used in the study.
}
\keyword{datasets}

